fix: remove the stripe form to a car-controll component and use it in carinfo
