+++ client/package.json
@@ -23,7 +23,6 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
-    "react-stripe-checkout": "^2.6.3",
     "swiper": "^8.4.4",
     "yup": "^0.32.11"
   },
