+++ client/src/components/StripForm/index.tsx
@@ -1,11 +1,14 @@
 import { Button } from '@mui/material';
 import { PaymentElement } from '@stripe/react-stripe-js';
 
+import './style.css';
+
 function StripeForm() {
   return (
-    <form>
+    <form className="stripe-form">
       <PaymentElement />
-      <Button type="submit">Pay</Button>
+      {' '}
+      <Button sx={{ color: '#0A20E6', borderColor: '#0A20E6' }} className="pay-button" type="submit">Pay</Button>
     </form>
   );
 }
