+++ client/src/pages/Car/index.tsx
@@ -1,24 +1,8 @@
-import { Button, Modal } from '@mui/material';
-import { Elements } from '@stripe/react-stripe-js';
-import { loadStripe } from '@stripe/stripe-js';
-import { useState } from 'react';
-
+import './style.css';
 import CarInfo from '../../components/CarInfoSection';
 import CarSlider from '../../components/CarSlider';
-import StripeForm from '../../components/StripForm';
-
-const stripePromise = loadStripe(
-  'pk_test_TYooMQauvdEDq54NiTphI7jx',
-);
 
 function Car() {
-  const [open, setOpen] = useState(false);
-  const handleOpen = () => setOpen(true);
-  const handleClose = () => setOpen(false);
-  const options = {
-    clientSecret: 'pi_1DseH42eZvKYlo2C5UQDyYph_secret_gowsU3j2SgDfFECrHNzE8UtGK',
-  };
-
   const testInfo = {
     id: 1,
     brand: 'Toyota',
@@ -51,17 +35,6 @@ function Car() {
     <>
       <CarSlider />
       <CarInfo carInfo={testInfo} />
-      <Button onClick={handleOpen}>Buy</Button>
-      <Modal
-        open={open}
-        onClose={handleClose}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description"
-      >
-        <Elements stripe={stripePromise} options={options}>
-          <StripeForm />
-        </Elements>
-      </Modal>
     </>
   );
 }
