+++ client/src/pages/Car/style.css
@@ -1,0 +1,11 @@
+.buttons-container {
+    margin: 2rem auto;
+  }
+  
+  .buttons-container button {
+    white-space: nowrap;
+    margin: 1rem;
+    width: 9rem;
+    height: 3rem;
+  }
+
